Minor API Restructuring
This change allows the API to be more extensible so it can handle more than just configuration and RPC functionality.
