--- madmin/api/modules/ftr_routecalc.py
@@ -1,5 +1,0 @@
-from .. import apiHandler
-
-class APIRouteCalc(apiHandler.ResourceHandler):
-    component = 'routecalc'
-    description = 'Add/Update/Delete routecalcs'
