+++ madmin/api/resources/ftr_devicesetting.py
@@ -1,6 +1,6 @@
-from .. import apiHandler
+from .resourceHandler import ResourceHandler
 
-class APIDeviceSetting(apiHandler.ResourceHandler):
+class APIDeviceSetting(ResourceHandler):
     component = 'devicepool'
     default_sort = 'devicepool'
     description = 'Add/Update/Delete Shared device settings'
