+++ madmin/api/resources/ftr_geofence.py
@@ -1,6 +1,6 @@
-from .. import apiHandler
+from .resourceHandler import ResourceHandler
 
-class APIGeofence(apiHandler.ResourceHandler):
+class APIGeofence(ResourceHandler):
     component = 'geofence'
     default_sort = 'name'
     description = 'Add/Update/Delete geofences'
