+++ madmin/api/resources/ftr_monlist.py
@@ -1,6 +1,6 @@
-from .. import apiHandler
+from .resourceHandler import ResourceHandler
 
-class APIMonList(apiHandler.ResourceHandler):
+class APIMonList(ResourceHandler):
     component = 'monivlist'
     default_sort = 'monlist'
     description = 'Add/Update/Delete Pokemon Lists'
