+++ madmin/api/resources/ftr_routecalc.py
@@ -1,0 +1,5 @@
+from .resourceHandler import ResourceHandler
+
+class APIRouteCalc(ResourceHandler):
+    component = 'routecalc'
+    description = 'Add/Update/Delete routecalcs'
