+++ madmin/api/resources/ftr_walker.py
@@ -1,6 +1,6 @@
-from .. import apiHandler
+from .resourceHandler import ResourceHandler
 
-class APIWalker(apiHandler.ResourceHandler):
+class APIWalker(ResourceHandler):
     component = 'walker'
     default_sort = 'walkername'
     description = 'Add/Update/Delete Walkers'
