+++ madmin/madmin.py
@@ -15,7 +15,7 @@ from madmin.routes.control import control
 from madmin.routes.map import map
 from madmin.routes.config import config
 from madmin.routes.path import path
-from madmin.api import APIHandler
+from madmin.api import APIEntry
 from madmin.reverseproxy import ReverseProxied
 
 
@@ -37,7 +37,7 @@ def madmin_start(args, db_wrapper: DbWrapper, ws_server, mapping_manager: Mappin
     statistics(db_wrapper, args, app, mapping_manager, data_manager)
     control(db_wrapper, args, mapping_manager, ws_server, logger, app, deviceUpdater)
     map(db_wrapper, args, mapping_manager, app, data_manager)
-    APIHandler(logger, app, data_manager, mapping_manager, ws_server, args.config_mode)
+    APIEntry(logger, app, data_manager, mapping_manager, ws_server, args.config_mode)
     config(db_wrapper, args, logger, app, mapping_manager, data_manager)
     path(db_wrapper, args, app, mapping_manager, jobstatus, data_manager)
